refactor: consolidate E2E tests for better maintainability
Consolidate 11 test files into 6 focused files while preserving all coverage:

NEW FILES:
- connection-lifecycle.spec.ts: Comprehensive connection testing (5 tests)
  • Basic connection, sequential reconnections, rapid reconnections
  • Explicit disconnect/reconnect cycles, multiple disconnect cycles
  • Preserves all reconnection scenarios from original tests

- session-management.spec.ts: Session ID and reuse testing (2 tests)
  • Session reuse verification, session ID rejection testing

IMPROVED FILES:
- uuid-format-compatibility.spec.ts: Now actually tests UUID formats (3 tests)
  • Tests short (9800), long (00009800-...), and mixed UUID formats
  • Properly overrides service config to test Noble transport handling

DELETED REDUNDANT FILES (7 files):
- real-device-session.spec.ts, zombie-reproduction.spec.ts
- session-pool-behavior.spec.ts, cleanup-state-integrity.spec.ts
- disconnect-reconnect-same-session.spec.ts, core-session-reuse.spec.ts
- session-rejection.spec.ts

Results:
- 45% reduction in test files (11 → 6)
- Maintained comprehensive coverage (22/22 tests pass)
- Cleaner organization by functionality
- All reconnection scenarios preserved
